Eliminate old_constants module
