+++ gym_sts/spaces/constants/combat.py
@@ -1,0 +1,287 @@
+import math
+
+
+MAX_ATTACK = 999
+MAX_ATTACK_TIMES = 15
+MAX_BLOCK = 999
+MAX_EFFECT = 999
+MAX_HAND_SIZE = 10
+MAX_NUM_ENEMIES = 6
+
+LOG_MAX_ATTACK = math.ceil(math.log(MAX_ATTACK, 2))
+LOG_MAX_ATTACK_TIMES = math.ceil(math.log(MAX_ATTACK_TIMES, 2))
+LOG_MAX_BLOCK = math.ceil(math.log(MAX_BLOCK, 2))
+LOG_MAX_EFFECT = math.ceil(math.log(MAX_EFFECT, 2))
+LOG_MAX_ENERGY = 4
+LOG_MAX_TURN = 8
+
+ALL_EFFECTS = [
+    "Conserve",
+    "Sharp Hide",
+    "Evolve",
+    "Double Tap",
+    "Draw Card",
+    "Demon Form",
+    "Thorns",
+    "Wraith Form v2",
+    "Angry",
+    "Time Warp",
+    "IntangiblePlayer",
+    "Ritual",
+    "Intangible",
+    "Flex",
+    "Thievery",
+    "Metallicize",
+    "Fading",
+    "BackAttack",
+    "Tools Of The Trade",
+    "Unawakened",
+    "Entangled",
+    "Plated Armor",
+    "Regenerate",
+    "Choked",
+    "Storm",
+    "Constricted",
+    "Feel No Pain",
+    "Dexterity",
+    "CorpseExplosionPower",
+    "Stasis",
+    "Panache",
+    "Magnetism",
+    "After Image",
+    "Nullify Attack",
+    "RechargingCore",
+    "GrowthPower",
+    "Flight",
+    "Rebound",
+    "Confusion",
+    "Mode Shift",
+    "Curiosity",
+    "Double Damage",
+    "Artifact",
+    "Berserk",
+    "Amplify",
+    "Anger",
+    "Loop",
+    "Creative AI",
+    "Next Turn Block",
+    "Generic Strength Up Power",
+    "Rage",
+    "Strength",
+    "Curl Up",
+    "Thousand Cuts",
+    "Equilibrium",
+    "Sadistic",
+    "Painful Stabs",
+    "Flame Barrier",
+    "Mayhem",
+    "Poison",
+    "Regeneration",
+    "Dark Embrace",
+    "Skill Burn",
+    "Weakened",
+    "Life Link",
+    "Draw Reduction",
+    "Juggernaut",
+    "Pen Nib",
+    "Electro",
+    "Fire Breathing",
+    "Buffer",
+    "Rupture",
+    "Malleable",
+    "Spore Cloud",
+    "Vulnerable",
+    "StrikeUp",
+    "Night Terror",
+    "Collect",
+    "BlockReturnPower",
+    "DevotionPower",
+    "EnergyDownPower",
+    "BattleHymn",
+    "WrathNextTurnPower",
+    "MasterRealityPower",
+    "Vault",
+    "DevaForm",
+    "Controlled",
+    "OmnisciencePower",
+    "PathToVictoryPower",
+    "Study",
+    "FreeAttackPower",
+    "OmegaPower",
+    "AngelForm",
+    "Adaptation",
+    "Nirvana",
+    "WireheadingPower",
+    "CannotChangeStancePower",
+    "NoSkills",
+    "LikeWaterPower",
+    "EndTurnDeath",
+    "EstablishmentPower",
+    "Vigor",
+    "WaveOfTheHandPower",
+    "Mantra",
+    "Lockon",
+    "Life Link",
+    "TheBomb",
+    "Compulsive",
+    "Hello",
+    "Winter",
+    "Split",
+    "Shifting",
+    "Focus",
+    "Phantasmal",
+    "Attack Burn",
+    "Minion",
+    "Noxious Fumes",
+    "Envenom",
+    "Brutality",
+    "NoBlockPower",
+    "Burst",
+    "EnergizedBlue",
+    "Explosive",
+    "Bias",
+    "DuplicationPower",
+    "Corruption",
+    "StaticDischarge",
+    "Slow",
+    "Lightning Mastery",
+    "Draw",
+    "Combust",
+    "Hex",
+    "Frail",
+    "Surrounded",
+    "Heatsink",
+    "TimeMazePower",
+    "DexLoss",
+    "Shackled",
+    "Retain Cards",
+    "Echo Form",
+    "Energized",
+    "Repair",
+    "No Draw",
+    "Invincible",
+    "FlickPower",
+    "AlwaysMad",
+    "HotHot",
+    "MasterRealityPower",
+    "FlowPower",
+    "DisciplinePower",
+    "DEPRECATEDCondense",
+    "EmotionalTurmoilPower",
+    "Grounded",
+    "Retribution",
+    "Serenity",
+    "Mastery",
+    "Barricade",
+    "Blur",
+    "BeatOfDeath",
+    "Infinite Blades",
+    "Accuracy",
+]
+# Wiki seems to list 108 buffs and debuffs, I may have missed a few
+NUM_EFFECTS = len(ALL_EFFECTS)
+
+ALL_INTENTS = [
+    "NONE",
+    "ATTACK",
+    "ATTACK_BUFF",
+    "ATTACK_DEBUFF",
+    "ATTACK_DEFEND",
+    "BUFF",
+    "DEBUFF",
+    "STRONG_DEBUFF",
+    "DEBUG",
+    "DEFEND",
+    "DEFEND_DEBUFF",
+    "DEFEND_BUFF",
+    "ESCAPE",
+    "MAGIC",
+    "SLEEP",
+    "STUN",
+    "UNKNOWN",
+]
+NUM_INTENTS = len(ALL_INTENTS)
+
+ALL_MONSTER_TYPES = [
+    "NONE",
+    "GremlinNob",
+    "GremlinTsundere",
+    "FungiBeast",
+    "GremlinThief",
+    "TheGuardian",
+    "FuzzyLouseNormal",
+    "GremlinWarrior",
+    "Looter",
+    "Lagavulin",
+    "AcidSlime_L",
+    "HexaghostOrb",
+    "Hexaghost",
+    "SlaverBlue",
+    "Sentry",
+    "AcidSlime_S",
+    "SpikeSlime_S",
+    "GremlinWizard",
+    "FuzzyLouseDefensive",
+    "SpikeSlime_M",
+    "AcidSlime_M",
+    "Cultist",
+    "Apology Slime",
+    "SlimeBoss",
+    "HexaghostBody",
+    "SpikeSlime_L",
+    "GremlinFat",
+    "SlaverRed",
+    "JawWorm",
+    "BronzeOrb",
+    "BookOfStabbing",
+    "TheCollector",
+    "Snecko",
+    "BanditBear",
+    "SlaverBoss",
+    "TorchHead",
+    "Shelled Parasite",
+    "Centurion",
+    "Chosen",
+    "BronzeAutomaton",
+    "Healer",
+    "BanditChild",
+    "BanditLeader",
+    "SphericGuardian",
+    "SnakePlant",
+    "Champ",
+    "Mugger",
+    "Byrd",
+    "GremlinLeader",
+    "Serpent",
+    "Darkling",
+    "Orb Walker",
+    "Donu",
+    "Maw",
+    "Spiker",
+    "AwakenedOne",
+    "TimeEater",
+    "Repulsor",
+    "WrithingMass",
+    "Deca",
+    "Exploder",
+    "Reptomancer",
+    "Transient",
+    "Nemesis",
+    "Dagger",
+    "GiantHead",
+    "SpireShield",
+    "SpireSpear",
+    "CorruptHeart",
+]
+NUM_MONSTER_TYPES = len(ALL_MONSTER_TYPES)
+
+ALL_ORBS = [
+    "NONE",  # Indicates the slot does not exist
+    "Empty",
+    "Dark",
+    "Frost",
+    "Lightning",
+    "Plasma",
+]
+NUM_ORBS = len(ALL_ORBS)
+MAX_ORB_SLOTS = 10
