+++ gym_sts/spaces/data.py
@@ -3,11 +3,12 @@ from pathlib import Path
 from typing import Dict, List, Tuple
 
 from gym_sts.constants import PROJECT_ROOT
-from gym_sts.spaces.old_constants import (
+from gym_sts.spaces.constants.events import (
     ALL_EVENTS,
     GLOBALLY_CHECKED_TEXTS,
     MAX_NUM_CUSTOM_TEXTS,
 )
+
 
 REMOVED_STRINGS = ["#r", "#y", "#g", "#b", "#p", "@", "~", "NL"]
