+++ gym_sts/spaces/observations/spaces.py
@@ -1,7 +1,8 @@
 from gym.spaces import Dict, Discrete, MultiBinary, MultiDiscrete, Tuple
 
+import gym_sts.spaces.constants.base as base_consts
 import gym_sts.spaces.constants.cards as card_consts
-from gym_sts.spaces import old_constants as constants
+import gym_sts.spaces.constants.combat as combat_consts
 
 
 def generate_card_space():
@@ -17,16 +18,16 @@ def generate_effect_space():
     effect_space = Dict(
         {
             "sign": Discrete(2),
-            "value": MultiBinary(constants.LOG_MAX_EFFECT),
+            "value": MultiBinary(combat_consts.LOG_MAX_EFFECT),
         }
     )
-    return Tuple([effect_space] * constants.NUM_EFFECTS)
+    return Tuple([effect_space] * combat_consts.NUM_EFFECTS)
 
 
 def generate_health_space():
     return Dict(
         {
-            "hp": MultiBinary(constants.LOG_MAX_HP),
-            "max_hp": MultiBinary(constants.LOG_MAX_HP),
+            "hp": MultiBinary(base_consts.LOG_MAX_HP),
+            "max_hp": MultiBinary(base_consts.LOG_MAX_HP),
         }
     )
